登陆时在会话中记录用户ID； 进入区服接口；

+++ Bean/GameServer/GameServerBean.cs
@@ -13,7 +13,7 @@
         /// <summary>
         /// 区服状态
         /// </summary>
-        public int RunStatus;
+        public byte RunStatus;
 
         /// <summary>
         /// 是否推荐
